01-Inventory-Management/01-Scraping-function/requirements_scraper.txt
requests
beautifulsoup4
lxml
pillow
pyairtable
python-dotenv
selenium
scrapy-splash
cloudscraper
scrapy 